+++ hibernate-core/src/main/java/org/hibernate/envers/configuration/internal/metadata/MetadataTools.java
@@ -136,7 +136,7 @@ public final class MetadataTools {
 	public static Element addColumn(
 			Element parent,
 			String name,
-			Integer length,
+			Long length,
 			Integer scale,
 			Integer precision,
 			String sqlType,
@@ -148,7 +148,7 @@ public final class MetadataTools {
 	public static Element addColumn(
 			Element parent,
 			String name,
-			Integer length,
+			Long length,
 			Integer scale,
 			Integer precision,
 			String sqlType,
